Move get_or_import_symbol onto Importer
