+++ crates/ruff/src/rules/flake8_simplify/rules/suppressible_exception.rs
@@ -7,7 +7,6 @@ use ruff_python_ast::call_path::compose_call_path;
 use ruff_python_ast::helpers;
 use ruff_python_ast::helpers::has_comments;
 
-use crate::autofix::actions::get_or_import_symbol;
 use crate::checkers::ast::Checker;
 use crate::registry::AsRule;
 
@@ -89,13 +88,11 @@ pub(crate) fn suppressible_exception(
 
             if fixable && checker.patch(diagnostic.kind.rule()) {
                 diagnostic.try_set_fix(|| {
-                    let (import_edit, binding) = get_or_import_symbol(
+                    let (import_edit, binding) = checker.importer.get_or_import_symbol(
                         "contextlib",
                         "suppress",
                         stmt.start(),
                         checker.semantic_model(),
-                        &checker.importer,
-                        checker.locator,
                     )?;
                     let replace_try = Edit::range_replacement(
                         format!("with {binding}({exception})"),
