+++ crates/ruff/src/rules/pylint/rules/sys_exit_alias.rs
@@ -3,7 +3,6 @@ use rustpython_parser::ast::{self, Expr, Ranged};
 use ruff_diagnostics::{AutofixKind, Diagnostic, Edit, Fix, Violation};
 use ruff_macros::{derive_message_formats, violation};
 
-use crate::autofix::actions::get_or_import_symbol;
 use crate::checkers::ast::Checker;
 use crate::registry::AsRule;
 
@@ -76,13 +75,11 @@ pub(crate) fn sys_exit_alias(checker: &mut Checker, func: &Expr) {
         );
         if checker.patch(diagnostic.kind.rule()) {
             diagnostic.try_set_fix(|| {
-                let (import_edit, binding) = get_or_import_symbol(
+                let (import_edit, binding) = checker.importer.get_or_import_symbol(
                     "sys",
                     "exit",
                     func.start(),
                     checker.semantic_model(),
-                    &checker.importer,
-                    checker.locator,
                 )?;
                 let reference_edit = Edit::range_replacement(binding, func.range());
                 #[allow(deprecated)]
