+++ crates/ruff/src/rules/pyupgrade/rules/lru_cache_with_maxsize_none.rs
@@ -4,7 +4,6 @@ use rustpython_parser::ast::{self, Constant, Expr, Keyword, Ranged};
 use ruff_diagnostics::{AlwaysAutofixableViolation, Diagnostic, Edit, Fix};
 use ruff_macros::{derive_message_formats, violation};
 
-use crate::autofix::actions::get_or_import_symbol;
 use crate::checkers::ast::Checker;
 use crate::registry::AsRule;
 
@@ -65,13 +64,11 @@ pub(crate) fn lru_cache_with_maxsize_none(checker: &mut Checker, decorator_list:
                 );
                 if checker.patch(diagnostic.kind.rule()) {
                     diagnostic.try_set_fix(|| {
-                        let (import_edit, binding) = get_or_import_symbol(
+                        let (import_edit, binding) = checker.importer.get_or_import_symbol(
                             "functools",
                             "cache",
                             expr.start(),
                             checker.semantic_model(),
-                            &checker.importer,
-                            checker.locator,
                         )?;
                         let reference_edit = Edit::range_replacement(binding, expr.range());
                         #[allow(deprecated)]
